@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2020 the QProv contributors.
+ * Copyright (c) 2023 the QProv contributors.
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -19,13 +19,24 @@
 
 package org.quantil.qprov.core.model;
 
-import javax.persistence.Entity;
-
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
 import org.openprovenance.prov.sql.Document;
 
-import lombok.Data;
-
+@EqualsAndHashCode(callSuper = true)
 @Data
 @Entity
 public class ProvTemplate extends Document {
+    @Id
+    private Long id;
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
 }
