@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2020 the QProv contributors.
+ * Copyright (c) 2023 the QProv contributors.
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -21,14 +21,15 @@ package org.quantil.qprov.core.repositories.prov;
 
 import java.util.Optional;
 
-import org.openprovenance.prov.sql.QualifiedName;
+import org.quantil.qprov.core.model.prov.ProvQualifiedName;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import org.springframework.stereotype.Repository;
 
 @RepositoryRestResource(exported = false)
 @Repository
-public interface QualifiedNameRepository extends JpaRepository<QualifiedName, Long> {
+public interface QualifiedNameRepository extends JpaRepository<ProvQualifiedName, Long> {
 
-    Optional<QualifiedName> findByUri(String uri);
+    Optional<ProvQualifiedName> findByUri(String uri);
 }
