@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2020 the QProv contributors.
+ * Copyright (c) 2023 the QProv contributors.
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -19,15 +19,11 @@
 
 package org.quantil.qprov.web.controller;
 
-import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.linkTo;
-import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.methodOn;
-
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 import java.util.stream.Stream;
-import javax.transaction.Transactional;
 import javax.ws.rs.QueryParam;
 
 import org.quantil.qprov.core.model.agents.Provider;
@@ -38,8 +34,13 @@ import org.quantil.qprov.core.repositories.ProviderRepository;
 import org.quantil.qprov.core.repositories.QPURepository;
 import org.quantil.qprov.web.Constants;
 import org.quantil.qprov.web.dtos.CalibrationMatrixDto;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.springframework.hateoas.CollectionModel;
 import org.springframework.hateoas.EntityModel;
 import org.springframework.hateoas.RepresentationModel;
@@ -52,10 +53,8 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import lombok.AllArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
+import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.linkTo;
+import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.methodOn;
 
 @io.swagger.v3.oas.annotations.tags.Tag(name = Constants.TAG_PROVIDER)
 @RestController
@@ -64,8 +63,7 @@ import lombok.extern.slf4j.Slf4j;
 @AllArgsConstructor
 @Slf4j
 public class AggregatedDataController {
-
-    private final Logger logger = LoggerFactory.getLogger(AggregatedDataController.class);
+    protected static final Logger logger = LogManager.getLogger();
 
     private final ProviderRepository providerRepository;
 
@@ -103,7 +101,6 @@ public class AggregatedDataController {
     @Operation(responses = {@ApiResponse(responseCode = "200"),
             @ApiResponse(responseCode = "404", description = "Provider or QPU not found or no calibration matrix available for this QPU.")})
     @GetMapping("/" + Constants.PATH_CALIBRATION_MATRIX)
-    @Transactional
     public ResponseEntity<CollectionModel<EntityModel<CalibrationMatrixDto>>> getCalibrationMatrix(@PathVariable UUID providerId,
                                                                                                    @PathVariable UUID qpuId,
                                                                                                    @QueryParam("latest") boolean latest) {
@@ -134,7 +131,7 @@ public class AggregatedDataController {
 
         final List<EntityModel<CalibrationMatrixDto>> entities = new ArrayList<>();
         calibrationMatrixStream.forEach(calibrationMatrix -> {
-            entities.add(new EntityModel<CalibrationMatrixDto>(CalibrationMatrixDto.createDTO(calibrationMatrix)));
+            entities.add(EntityModel.of(CalibrationMatrixDto.createDTO(calibrationMatrix)));
         });
         logger.debug("Retrieved {} calibration matrix records for QPU with name: {}", entities.size(), qpu.getName());
 
@@ -142,6 +139,6 @@ public class AggregatedDataController {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
         }
 
-        return ResponseEntity.ok(new CollectionModel<>(entities));
+        return ResponseEntity.ok(CollectionModel.of(entities));
     }
 }
