@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2020 the QProv contributors.
+ * Copyright (c) 2023 the QProv contributors.
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -38,6 +38,11 @@ import org.quantil.qprov.core.repositories.QPURepository;
 import org.quantil.qprov.core.repositories.QubitRepository;
 import org.quantil.qprov.web.Constants;
 import org.quantil.qprov.web.dtos.GateCharacteristicsDto;
+
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.hateoas.CollectionModel;
 import org.springframework.hateoas.EntityModel;
 import org.springframework.http.HttpStatus;
@@ -47,11 +52,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import lombok.AllArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 
 @io.swagger.v3.oas.annotations.tags.Tag(name = Constants.TAG_PROVIDER)
 @RestController
@@ -120,13 +120,13 @@ public class GateCharacteristicsController {
 
         final List<EntityModel<GateCharacteristicsDto>> entities = new ArrayList<>();
         gateCharacteristicsStream.forEach(gateCharacteristic -> {
-            entities.add(new EntityModel<GateCharacteristicsDto>(GateCharacteristicsDto.createDTO(gateCharacteristic)));
+            entities.add(EntityModel.of(GateCharacteristicsDto.createDTO(gateCharacteristic)));
         });
 
         if (entities.isEmpty()) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
         }
 
-        return ResponseEntity.ok(new CollectionModel<>(entities));
+        return ResponseEntity.ok(CollectionModel.of(entities));
     }
 }
