@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2020 the QProv contributors.
+ * Copyright (c) 2023 the QProv contributors.
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -19,9 +19,6 @@
 
 package org.quantil.qprov.web.controller;
 
-import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.linkTo;
-import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.methodOn;
-
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -35,8 +32,13 @@ import org.quantil.qprov.core.repositories.QPURepository;
 import org.quantil.qprov.core.repositories.QubitRepository;
 import org.quantil.qprov.web.Constants;
 import org.quantil.qprov.web.dtos.QubitDto;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.springframework.hateoas.CollectionModel;
 import org.springframework.hateoas.EntityModel;
 import org.springframework.hateoas.Link;
@@ -48,10 +50,8 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import lombok.AllArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
+import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.linkTo;
+import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.methodOn;
 
 @io.swagger.v3.oas.annotations.tags.Tag(name = Constants.TAG_PROVIDER)
 @RestController
@@ -60,8 +60,7 @@ import lombok.extern.slf4j.Slf4j;
 @AllArgsConstructor
 @Slf4j
 public class QubitController {
-
-    private static final Logger logger = LoggerFactory.getLogger(QubitController.class);
+    protected static final Logger logger = LogManager.getLogger();
 
     private final ProviderRepository providerRepository;
 
@@ -100,7 +99,7 @@ public class QubitController {
                 }
         );
 
-        final var collectionModel = new CollectionModel<>(qubitEntities);
+        final var collectionModel = CollectionModel.of(qubitEntities);
         collectionModel.add(qubitLinks);
         collectionModel.add(linkTo(methodOn(QubitController.class).getQubits(providerId, qpuId)).withSelfRel());
         return ResponseEntity.ok(collectionModel);
@@ -137,7 +136,7 @@ public class QubitController {
     }
 
     private EntityModel<QubitDto> createQubitDto(UUID providerId, UUID qpuId, Qubit qubit) {
-        final EntityModel<QubitDto> qpuDto = new EntityModel<QubitDto>(QubitDto.createDTO(qubit));
+        final EntityModel<QubitDto> qpuDto = EntityModel.of(QubitDto.createDTO(qubit));
         qpuDto.add(linkTo(methodOn(QubitController.class).getQubit(providerId, qpuId, qubit.getDatabaseId()))
                 .withSelfRel());
         qpuDto.add(linkTo(methodOn(QubitCharacteristicsController.class).getQubitCharacterisitcs(providerId, qpuId, qubit.getDatabaseId(), false))
