@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2020 the QProv contributors.
+ * Copyright (c) 2023 the QProv contributors.
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -21,12 +21,11 @@ package org.quantil.qprov.web.dtos;
 
 import java.util.Map;
 
-import org.openprovenance.prov.model.Namespace;
-
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
+import org.openprovenance.prov.model.Namespace;
 
 @EqualsAndHashCode
 @Data
